requirements.txt
dotenv==0.9.9
pillow==12.0.0
numpy==2.2.6
matplotlib==3.10.7
tqdm==4.67.1
mlflow==3.6.0
segmentation-models-pytorch==0.5.0
opencv-python==4.12.0.88
albumentations==2.0.8
ipykernel==7.1.0
rich==14.2.0